Switch vectors to float32 end to end and speed up the cosine scan
Embeddings were float64 in memory and float32 on disk: the SQLite store
quantized on write, and every query converted the corpus back to
float64 one vector at a time. Now the embedder returns float32, the
chunk and the record carry float32, and scoring runs float32 — the
stored bytes are the scored bytes, memory halves, and the per-query
conversion disappears.

Cosine32 keeps float64 accumulators but widens them in blocks of four
lanes: a single accumulator serializes on its own add, which profiling
showed was the real bound on the scan, not memory bandwidth. Rankings
against the old float64 Cosine were checked over 500 randomized
trials — no top-5 disagreement — and the SQLite blob format is
unchanged, so existing files open as they are.

Brute-force scan costs, 1536 dimensions, before -> after:
- 200k chunks: 591ms -> 297ms
- 50k chunks: 143ms -> 81ms
- 10k chunks: 27ms -> 18ms

+++ internal/cli/query.go
@@ -90,7 +90,7 @@ func filtered(st store.Store, src *string) ([]store.Record, error) {
 }
 
 // embedOne runs the embedder over the single query text.
-func embedOne(emb embed.Embedder, query string) ([]float64, error) {
+func embedOne(emb embed.Embedder, query string) ([]float32, error) {
 	vecs, err := emb.Embed([]string{query})
 	if err != nil {
 		return nil, err
